Show 'installations' in menu if user is admin
