+++ web-app/src/components/Header/index.js
@@ -1,33 +1,13 @@
-/* global window */
-import React, { Component } from 'react';
-import './Header.css';
+import { connect } from 'react-redux';
 
-import logo from '../../theme/crest.png';
+import Header from './Header';
+import { isAdmin } from '../../selectors/application';
 
-import api, { removeToken } from '../../utils/api';
+const mapStateToProps = (state) => ({
+  isAdmin: isAdmin(state),
+});
 
-class Header extends Component {
-  logout() {
-    api
-      .post('/auth/logout')
-      .then(() => {
-        removeToken();
-        window.location = '/login';
-      });
-  }
-
-  render() {
-    return (
-      <div className="Header">
-        <div className="Header__Wrapper">
-          <img src={logo} alt="Logo" className="Header__Logo" />
-          <div className="Logout" onClick={() => this.logout()} >
-            Logout
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-export default Header;
+export default connect(
+  mapStateToProps,
+  undefined,
+)(Header);
